+++ src/Analyzers/CSharp/Analysis/UnnecessarySemicolonAtEndOfDeclarationAnalyzer.cs
@@ -9,7 +9,7 @@ using Microsoft.CodeAnalysis.Diagnostics;
 namespace Roslynator.CSharp.Analysis;
 
 [DiagnosticAnalyzer(LanguageNames.CSharp)]
-public sealed class AvoidSemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAnalyzer
+public sealed class UnnecessarySemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAnalyzer
 {
     private static ImmutableArray<DiagnosticDescriptor> _supportedDiagnostics;
 
@@ -18,7 +18,7 @@ public sealed class AvoidSemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAna
         get
         {
             if (_supportedDiagnostics.IsDefault)
-                Immutable.InterlockedInitialize(ref _supportedDiagnostics, DiagnosticRules.AvoidSemicolonAtEndOfDeclaration);
+                Immutable.InterlockedInitialize(ref _supportedDiagnostics, DiagnosticRules.UnnecessarySemicolonAtEndOfDeclaration);
 
             return _supportedDiagnostics;
         }
@@ -45,7 +45,7 @@ public sealed class AvoidSemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAna
         if (semicolon.Parent is not null
             && !semicolon.IsMissing)
         {
-            DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.AvoidSemicolonAtEndOfDeclaration, semicolon);
+            DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.UnnecessarySemicolonAtEndOfDeclaration, semicolon);
         }
     }
 
@@ -53,12 +53,15 @@ public sealed class AvoidSemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAna
     {
         var declaration = (ClassDeclarationSyntax)context.Node;
 
-        SyntaxToken semicolon = declaration.SemicolonToken;
-
-        if (semicolon.Parent is not null
-            && !semicolon.IsMissing)
+        if (declaration.CloseBraceToken.IsKind(SyntaxKind.CloseBraceToken))
         {
-            DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.AvoidSemicolonAtEndOfDeclaration, semicolon);
+            SyntaxToken semicolon = declaration.SemicolonToken;
+
+            if (semicolon.Parent is not null
+                && !semicolon.IsMissing)
+            {
+                DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.UnnecessarySemicolonAtEndOfDeclaration, semicolon);
+            }
         }
     }
 
@@ -66,12 +69,15 @@ public sealed class AvoidSemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAna
     {
         var declaration = (InterfaceDeclarationSyntax)context.Node;
 
-        SyntaxToken semicolon = declaration.SemicolonToken;
-
-        if (semicolon.Parent is not null
-            && !semicolon.IsMissing)
+        if (declaration.CloseBraceToken.IsKind(SyntaxKind.CloseBraceToken))
         {
-            DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.AvoidSemicolonAtEndOfDeclaration, semicolon);
+            SyntaxToken semicolon = declaration.SemicolonToken;
+
+            if (semicolon.Parent is not null
+                && !semicolon.IsMissing)
+            {
+                DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.UnnecessarySemicolonAtEndOfDeclaration, semicolon);
+            }
         }
     }
 
@@ -79,12 +85,15 @@ public sealed class AvoidSemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAna
     {
         var declaration = (StructDeclarationSyntax)context.Node;
 
-        SyntaxToken semicolon = declaration.SemicolonToken;
-
-        if (semicolon.Parent is not null
-            && !semicolon.IsMissing)
+        if (declaration.CloseBraceToken.IsKind(SyntaxKind.CloseBraceToken))
         {
-            DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.AvoidSemicolonAtEndOfDeclaration, semicolon);
+            SyntaxToken semicolon = declaration.SemicolonToken;
+
+            if (semicolon.Parent is not null
+                && !semicolon.IsMissing)
+            {
+                DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.UnnecessarySemicolonAtEndOfDeclaration, semicolon);
+            }
         }
     }
 
@@ -99,7 +108,7 @@ public sealed class AvoidSemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAna
             if (semicolon.Parent is not null
                 && !semicolon.IsMissing)
             {
-                DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.AvoidSemicolonAtEndOfDeclaration, semicolon);
+                DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.UnnecessarySemicolonAtEndOfDeclaration, semicolon);
             }
         }
     }
@@ -113,7 +122,7 @@ public sealed class AvoidSemicolonAtEndOfDeclarationAnalyzer : BaseDiagnosticAna
         if (semicolon.Parent is not null
             && !semicolon.IsMissing)
         {
-            DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.AvoidSemicolonAtEndOfDeclaration, semicolon);
+            DiagnosticHelpers.ReportDiagnostic(context, DiagnosticRules.UnnecessarySemicolonAtEndOfDeclaration, semicolon);
         }
     }
 }